// apps/console-api/src/gen/hue/v1/lighting.ts
export interface LightPut {
  type?: string | undefined;
  on?: On | undefined;
  dimming?: Dimming | undefined;
  dimmingDelta?: DimmingDelta | undefined;
  colorTemperatureDelta?: ColorTemperatureDelta | undefined;
  dynamics?: LightDynamics | undefined;
  alert?: Alert | undefined;
  signaling?: Signaling | undefined;
  mode?: LightPut_Mode | undefined;
  gradient?: Gradient | undefined;
  effects?: Effects | undefined;
  effectsV2?: EffectsV2 | undefined;
  timedEffects?: LightPut_TimedEffects | undefined;
  powerup?: Powerup | undefined;
  colour:
    | { $case: "color"; color: Color }
    | { $case: "colorTemperature"; colorTemperature: ColorTemperature }
    | undefined;
}

export enum LightPut_Mode {
  MODE_UNSPECIFIED = 0,
  MODE_NORMAL = 1,
  MODE_STREAMING = 2,
  UNRECOGNIZED = -1,
}
// ...
export interface LightPut_TimedEffects {
  effect?: SupportedTimedEffects | undefined;
  duration?: number | undefined;
}
// ...
export const LightPut_TimedEffects: MessageFns<LightPut_TimedEffects> = {
  encode(message: LightPut_TimedEffects, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.effect !== undefined) {
      writer.uint32(8).int32(message.effect);
    }
    if (message.duration !== undefined) {
      writer.uint32(16).int32(message.duration);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): LightPut_TimedEffects {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const previousRecursionDepth = (reader as any).__tsProtoDecodeDepth ?? 0;
    if (previousRecursionDepth >= 100) {
      throw new globalThis.Error("protobuf decode recursion limit exceeded");
    }
    (reader as any).__tsProtoDecodeDepth = previousRecursionDepth + 1;
    try {
      const end = length === undefined ? reader.len : reader.pos + length;
      const message = createBaseLightPut_TimedEffects();
      while (reader.pos < end) {
        const tag = reader.uint32();
        switch (tag >>> 3) {
          case 1: {
            if (tag !== 8) {
              break;
            }

            message.effect = reader.int32() as any;
            continue;
          }
          case 2: {
            if (tag !== 16) {
              break;
            }

            message.duration = reader.int32();
            continue;
          }
        }
        if ((tag & 7) === 4 || tag === 0) {
          break;
        }
        reader.skip(tag & 7);
      }
      return message;
    } finally {
      (reader as any).__tsProtoDecodeDepth = previousRecursionDepth;
    }
  },

  fromJSON(object: any): LightPut_TimedEffects {
    return {
      effect: isSet(object.effect) ? supportedTimedEffectsFromJSON(object.effect) : undefined,
      duration: isSet(object.duration) ? globalThis.Number(object.duration) : undefined,
    };
  },

  toJSON(message: LightPut_TimedEffects): unknown {
    const obj: any = {};
    if (message.effect !== undefined) {
      obj.effect = supportedTimedEffectsToJSON(message.effect);
    }
    if (message.duration !== undefined) {
      obj.duration = Math.round(message.duration);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<LightPut_TimedEffects>, I>>(base?: I): LightPut_TimedEffects {
    return LightPut_TimedEffects.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<LightPut_TimedEffects>, I>>(object: I): LightPut_TimedEffects {
    const message = createBaseLightPut_TimedEffects();
    message.effect = object.effect ?? undefined;
    message.duration = object.duration ?? undefined;
    return message;
  },
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends { $case: string } ? { [K in keyof Omit<T, "$case">]?: DeepPartial<T[K]> } & { $case: T["$case"] }
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}